requirements.txt
flask
flask-cors
flask-sqlalchemy
flask-jwt-extended
mysql-connector-python
bcrypt
python-dotenv

pytesseract 
pillow 
flask
PyPDF2 
pdf2image
openai 
langchain 
pandas
langchain_community
openpyxl
sentence_transformers
faiss-cpu
mysql-connector-python
sqlalchemy
pymysql 
requests 
pillow
flask-login
langchain_openai
transformers
torch
langchain-groq
selenium
webdriver-manager

